+++ packages/nimble-components/src/table-column/anchor/cell-view/template.ts
@@ -26,7 +26,7 @@ export const template = html<TableColumnAnchorCellView>`
                 target="${x => x.columnConfig?.target}"
                 type="${x => x.columnConfig?.type}"
                 download="${x => x.columnConfig?.download}"
-                underline-hidden="${x => x.columnConfig?.underlineHidden}"
+                ?underline-hidden="${x => x.columnConfig?.underlineHidden}"
                 appearance="${x => x.columnConfig?.appearance}"
                 title=${x => (x.hasOverflow ? x.text : null)}
             >